refactor: rename model file name
